feat: created component hierarchy
Having drawn all of the views of the app and broken them down into a hierarchy of components, these components have been created and linked together through import and export.
On this commit the components are empty, with placeholders.

react-router has been installed and paths have been specified in App.js for some components.